Fix clippy lints: range loop in PIN_BUFS parser, unnecessary cast in test
CI's stable clippy (1.96) flags the indexed range loop in the
MEGANEURA_PIN_BUFS debug parser; use slice fill. Also drop a redundant
u32 cast in inference_parity_large.

+++ tests/inference_parity_large.rs
@@ -96,7 +96,7 @@ fn group_norm_silu_parity_sr_scale() {
     let x = g.input("x", &[n]);
     let w = g.parameter("w", &[c as usize]);
     let b = g.parameter("b", &[c as usize]);
-    let y = g.group_norm(x, w, b, 1, c, (hw * hw) as u32, groups, 1e-5);
+    let y = g.group_norm(x, w, b, 1, c, hw * hw, groups, 1e-5);
     let y = g.silu(y);
     g.set_outputs(vec![y]);
